+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 
 - `pairs-to-llama2` filter now has an optional `--prefix` parameter to use with the instruction
 - added the `pretrain-sentences-to-pairs` filter for generating artificial prompt/response datasets from pretrain data
+- requires seppl>=0.0.9 now
 
 
 0.0.3 (2023-11-10)