+++ CHANGES.rst
@@ -11,6 +11,7 @@ Changelog
   process (e.g., used when generating help screens for derived libraries that shouldn't output the
   base plugins as well)
 - `llm-registry` and `llm-help` now allow specifying exlcuded modules via `-e/--excluded_modules` option
+- `to-alpaca` writer now has the `-a/--ensure_ascii` flag to enforce ASCII compatibility in the output
 
 
 0.0.3 (2023-11-10)