+++ CHANGES.rst
@@ -12,6 +12,8 @@ Changelog
   base plugins as well)
 - `llm-registry` and `llm-help` now allow specifying exlcuded modules via `-e/--excluded_modules` option
 - `to-alpaca` writer now has the `-a/--ensure_ascii` flag to enforce ASCII compatibility in the output
+- added global option `-u/--update_interval` to `convert` tool to customize how often progress of # records
+  processed is being output in the console (default: 1000)
 
 
 0.0.3 (2023-11-10)